Enable back navigation using physical buttons
This especially helps to make the Android back button work.
If we cannot go back further, we let the platform decide whether
to quit or to do nothing.

It also reacts on Alt+Left shortcut but not the backspace key, since
this can cause  unwanted navigation when typing in a textfield (ie.
mashing the backspace key to clear it)

Change-Id: I4c0c01285628d8f810a781c79ffb3fdee0489b8e
Reviewed-by: Caroline Chao <caroline.chao@digia.com>