Review LongTermPage to improve performance on mobile devices
The Separator has its implicitWidth and implicitHeight
changing when the screen rotates. This causes the layout
to be redrawn. Its implicitWidth is only used as margin, thus
rewriting the LongTermPage to use left/right margins instead of
left/right Separator as part of the layout.

The ApplicationInfo.isMobile property was not defined, which caused
an additional and unnecessary binding on page width/height changes.
When the screen rotates the width and height change, thus the
physicalSizeChanged signal is emitted twice. Use primaryOrientationChanged
instead now.

Those changes improve the performance when redrawing the screen while
rotating a device.

Change-Id: I09d3b370d45e7bb62ba20dd1fc7182747c468225
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>