Fix image provider URL loading
The local file URL is passed in with the file scheme removed. This means
that URL encodings to the local path may end up in the file name, which
may not load on some platforms. So, reparse this as a file URL and
convert it to a local file name string.

Change-Id: Iac32e7adde717e9ab2dce77d9237cd1cfc526fe6
Reviewed-by: Caroline Chao <caroline.chao@digia.com>