Make delete button also available for accessibility
Previously the gridlayout was the button, but then it would ignore
all its children (including the delete button)

Now we move the accessible button to the MouseArea, which is not
a parent of the delete button. This also ensures that the geometry of
the button is the same for both accessible and non-accessible users.

Change-Id: I139bee8836407ed3d8b6da08249eae351d7ecb56
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@theqtcompany.com>